Set ETS table heir, so ext_mod process inherits the table
An ETS table is automatically destroyed when its parent process dies.
When mod_cron is installed using the module_install command,
the calling process dies immediately, and ETS table gets destroyed.

That problem didn't appear when the module is already installed
and configured when ejabberd starts: in that case the calling process
is ext_mod, which is always alive.